@@ -2,11 +2,11 @@
 
   <Import Project="..\..\core.props" />
   <PropertyGroup>
-    <TargetFrameworks>netstandard1.0;netstandard2.0;netstandard2.1;net5.0;net6.0</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;netstandard2.1;net6.0;net7.0</TargetFrameworks>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json.Schema" Version="3.*" Condition="'$(TargetFramework)' != 'netstandard1.0'" />
+    <PackageReference Include="Newtonsoft.Json.Schema" Version="3.*" />
   </ItemGroup>
 
 </Project>
